refactor: support for shared widget item
