+++ src/components/UserCredentialItem/index.ts
@@ -1,0 +1,3 @@
+import UserCredentialItem from './UserCredentialItem'
+
+export default UserCredentialItem
