--- src/modules/UserChallenges/Item/index.ts
@@ -1,3 +1,0 @@
-import UserChallengeItem from './UserChallengeItem'
-
-export default UserChallengeItem
